# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Movie Night",
        "version": "0.1"
    },
    "paths": {
        "/api/movies": {
            "get": {
                "tags": [
                    "movies"
                ],
                "summary": "Find all movies",
                "description": "Returns a list of movies",
                "operationId": "App\\Http\\Controllers\\MovieController::getMovies",
                "responses": {
                    "200": {
                        "description": "A list of movies"
                    }
                }
            }
        },
        "/api/movies/{id}": {
            "get": {
                "tags": [
                    "movies"
                ],
                "summary": "Find movie by ID",
                "description": "Returns a single movie",
                "operationId": "getMovie",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Movie ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "404": {
                        "description": "Movie not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "movies"
                ],
                "operationId": "App\\Http\\Controllers\\MovieController::createMovie",
                "responses": {
                    "200": {
                        "description": "Movie added to database"
                    }
                }
            }
        },
        "/api/lists/{listId}": {
            "post": {
                "tags": [
                    "lists"
                ],
                "summary": "Adds a movie to a list",
                "operationId": "addToList",
                "parameters": [
                    {
                        "name": "listId",
                        "in": "path",
                        "description": "ID of list in which to place the movie",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Movie to be added",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Title of the movie",
                                        "type": "string"
                                    },
                                    "year": {
                                        "description": "Year the movie was made",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Added to list"
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}